Include inclusive namespaces when canonicalizing signed info
I ran into this bug while trying to sign a request for an old SOAP API.
Looks like the SignedInfo node can have a child that looks like this:

```xml
<ds:CanonicalizationMethod Algorithm="http://www.w3.org/2001/10/xml-exc-c14n#">
  <ec:InclusiveNamespaces xmlns:ec="http://www.w3.org/2001/10/xml-exc-c14n#" PrefixList="foo"/>
</ds:CanonicalizationMethod>
```

When canonicalizing the signed info to actually create the signature,
the `foo` namespace needs to be included. This commit adds that
behavior.